LibTerraria+Server: Implement PlayerDead packet, don't sync certain packets while the client is still connecting, and be sure to sync existing clients to the newly connecting one.
My previous commit mentions this, that we need to send a PlayerActive
packet to reset the stale player for existing clients, but needed to
delay the sending of certain things before we could do that. So now we
do! This implements Client::full_sync to sync everything about a client
to another. We call this once a client has finished connecting to tell
them of everyone else, and everyone else of them.