LibTerraria+Server: Support `wld` file loading, and make Server use `test.wld` by default! Hooray!
There's still quite a bit of optimization to do for the TileMap and
World class. We also don't save modifictaions at all, or sent the proper
WorldData packet. We are slowly getting there...

I'm aware of a certain world that will not sent with TileSection
packets... not sure why yet.