Terraria: Improve World loading semantics
A World _always_ needs a `TileMap`, so it's now a `NonnullRefPtr`.

The `try_parse_world` result was changed to return a `NonnullRefPtr` on
success, as if should only fail if an error-result is returned.

We now also try to create the World instance as late as possible,
reading the version, file metadata, and header before even trying to.

Instead of failing unexpectedly and ambigiously on world versions that
we don't support, we now require world versions to be the exact one we
support.