Meta: Fix compilation of serenity, use `ninja` during GitHub CI
_Something_ changed during updating the submodule that caused the build
to fail, that I hadn't noticed because the missing library was already
there and never cleaned up from older builds. Instead of winging it,
let's use something tried and tested.

`cmake/FetchLagom.cmake` is originally from `linusg/libjs-test262`,
written by Linus, David, and Andrew. Thanks to Andrew for the suggestion
of using it!

To speed up our CI, and align it with the recommended build process, we
now use ninja during GitHub CI.